+++ cmd/hlc/daemon/install_hlc.sh
@@ -6,7 +6,7 @@ echo "
 Description=HiddenLakeComposite
 
 [Service]
-ExecStart=/root/hlc_amd64_linux --path /root
+ExecStart=/usr/local/bin/hlc_amd64_linux --path ~/.config/hidden-lake
 Restart=always
 RestartSec=10
 
@@ -14,7 +14,7 @@ RestartSec=10
 WantedBy=multi-user.target
 " > /etc/systemd/system/hidden_lake_composite.service
 
-cd /root && \
+cd /usr/local/bin && \
     rm -f hlc_amd64_linux && \
     wget https://github.com/number571/hidden-lake/releases/latest/download/hlc_amd64_linux && \
     chmod +x hlc_amd64_linux