+++ internal/composite/pkg/app/app_test.go
@@ -127,8 +127,4 @@ func TestInitApp(t *testing.T) {
 	if _, err := InitApp([]string{"--path", tcTestdataPath}, tgFlags); err != nil {
 		t.Fatal(err)
 	}
-
-	if _, err := InitApp([]string{"--path", "./not_exist/path/to/hlc"}, tgFlags); err == nil {
-		t.Fatal("success init app with undefined dir key")
-	}
 }