+++ internal/services/pinger/pkg/app/errors.go
@@ -18,4 +18,5 @@ var (
 	ErrClose      = &SAppError{"close"}
 	ErrInitConfig = &SAppError{"init config"}
 	ErrSetBuild   = &SAppError{"set build"}
+	ErrMkdirPath  = &SAppError{"mkdir path"}
 )